requirements.txt
pip-licenses
pyyaml